The player purge will be disabled if the days are equals to or less than 0

+++ src/net/nifheim/beelzebu/coins/core/database/MySQL.java
@@ -155,7 +155,7 @@ public class MySQL implements Database {
                     st.executeUpdate("ALTER TABLE `" + prefix + "Multipliers` DROP COLUMN starttime;");
                 }
                 core.debug("The multipliers table was updated");
-                if (core.getConfig().getBoolean("General.Purge.Enabled", true)) {
+                if (core.getConfig().getBoolean("General.Purge.Enabled", true) && core.getConfig().getInt("General.Purge.Days") > 0) {
                     st.executeUpdate("DELETE FROM " + prefix + "Data WHERE lastlogin < " + (System.currentTimeMillis() - (core.getConfig().getInt("General.Purge.Days", 60) * 86400000L)) + ";");
                     core.debug("Inactive users were removed from the database.");
                 }
